docs/requirements.txt
Django>=4.2.0,<5.0
markdown>=3.4.0
Pygments>=2.14.0
gunicorn>=21.0.0
whitenoise>=6.0.0
python-dotenv>=1.0.0
dj-database-url>=2.0.0